Huge refactor.
All public fields have been changed to private.

Cube, Cover and CubeSet must be located in the same package. Cube has a key that locks inputStates. When a Cube is added to a Cover the key is locked. While the key is locked, the only way to change InputStates of that cube is to use Cover#changeCubeInputState. This is a safety measure so column count data doesn't get corrupted by accident.

This mechanism relies on package-local modifier.
Note: CubeSet is now package-local.