CubeArray no longer extends ArrayList. Now it manages an internal ArrayList (composition over inheritance).
CubeArray now enforces that a Cube can belong to only one CubeArray at the same time. If someone tries to break this contract an appropriate exception will arise. Also since CubeArray is a composition it is guaranteed that all cubes that are being added will go through these kinds of maintenance.

Add several tests to make sure the CubeArray and Cover work properly (that bitCount values are calculated correctly).