Add equals method to Cover for debug purposes.
Refactor Cover#ShannonCofactors to no longer accept arbitrary cubes. Instead it requires an index for the input part of the cover and then it creates an appropriate cube for the job.

Add a correct way to complement a cube according to De Morgan's laws.

Add several tests.

Fix some bugs.