Add several ways to construct CubeArrays and Covers.
Add several ways to change input and output states to Cubes. All of these are convenience methods.

Finish debugging single output complementation.

Implement multiple output complementation. Unsure how to test if the methods is working. Current test relies on checking if intersection of original and complement is empty and if union of original and complement is a tautology.